+++ BetterChests/Framework/UI/Menus/BaseMenu.cs
@@ -1,0 +1,6 @@
+namespace StardewMods.BetterChests.Framework.UI.Menus;
+
+using StardewValley.Menus;
+
+/// <summary>Base menu.</summary>
+internal abstract class BaseMenu : IClickableMenu { }