+++ BetterChests/Framework/UI/Menus/CategorizeMenu.cs
@@ -1,0 +1,4 @@
+namespace StardewMods.BetterChests.Framework.UI.Menus;
+
+/// <summary>Menu for categorizing storages.</summary>
+internal sealed class CategorizeMenu : BaseMenu { }