+++ BetterChests/Framework/UI/Menus/ConfigMenu.cs
@@ -1,0 +1,7 @@
+namespace StardewMods.BetterChests.Framework.UI.Menus;
+
+/// <summary>The main config options for Better Chests.</summary>
+internal sealed class ConfigMenu : BaseMenu
+{
+    // Tabbed menu
+}