+++ BetterChests/Framework/UI/Menus/LayoutMenu.cs
@@ -1,0 +1,4 @@
+namespace StardewMods.BetterChests.Framework.UI.Menus;
+
+/// <summary>Menu for configuring layout.</summary>
+internal sealed class LayoutMenu : BaseMenu { }