+++ BetterChests/Framework/UI/Menus/SortMenu.cs
@@ -1,0 +1,4 @@
+namespace StardewMods.BetterChests.Framework.UI.Menus;
+
+/// <summary>Menu for sorting options.</summary>
+internal sealed class SortMenu : BaseMenu { }