+++ BetterChests/Framework/UI/Menus/TabMenu.cs
@@ -1,0 +1,4 @@
+namespace StardewMods.BetterChests.Framework.UI.Menus;
+
+/// <summary>Menu for customizing tabs.</summary>
+internal sealed class TabMenu : BaseMenu { }